requirement.txt
fastapi==0.95.2
uvicorn==0.23.2
bcrypt==4.0.1
PyJWT==2.8.0
